More cleanup to remove as much IOTC from renesas files as possible
